+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### StandaloneThirdPerson v1.3.0
+- Fixed issue with mod not showing when not launching through steam (--no-vr)
+- Added secondary keybinds for third person toggle and freeform toggle, you can leave these as `None` or set them to `LeftControl` for example
+- Added a setting to enable/disable freeform
+
 ### StandaloneThirdPerson v1.2.1
 - Added a few more null checks to help track down annoying bug
 - Keybind value case no longer matters (`T` and `t` are now the same and both are valid)